comit
implementação
