+++ projetoIntegrador/Assets/Jogos/JogoDaColeta/Scripts/GameControllerJC.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: fd9fb1e1bf6ee6a4c89cab83ca8008d9
+guid: de6a7e35fbfc5bd4aa441316c6c564b6
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
