--- projetoIntegrador/Assets/Jogos/JogoDaColeta/Scripts/ObjectGenerator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 66d7961770bf62c499d60a2162ee7491
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
