+++ projetoIntegrador/Assets/Jogos/JogoDaColeta/Scripts/ObjectsFalling.cs
@@ -6,21 +6,6 @@ using TMPro;
 public class ObjectsFalling : MonoBehaviour
 
 {
-    public static int points = 0;
-
-    void OnTriggerEnter2D(Collider2D other)
-    {
-        if (other.CompareTag("Player"))
-        {
-            // Destruir o objeto que caiu
-            Destroy(gameObject);
-
-            // Aumentar a pontuação do jogador
-            points++;
-            Debug.Log("pontos:" + points);
-        }
-    }
-
     void Update()
     {
         if (transform.position.y < -6f)
