+++ projetoIntegrador/Assets/Jogos/JogoDaColeta/Scripts/PlayerColeta.cs
@@ -2,11 +2,25 @@ using System.Collections;
 using System.Collections.Generic;
 using TMPro;
 using UnityEngine;
+using UnityEngine.UIElements;
 
 public class PlayerColeta : MonoBehaviour
 {
     public float velocidade = 5f;
-    static int missingObjects = 0; 
+    static int missingObjects = 0;
+    public int points = 0;
+
+    void OnTriggerEnter2D(Collider2D other)
+    {
+        if (other.CompareTag("Player"))
+        {
+            // Destruir o objeto que caiu
+            Destroy(gameObject);
+
+            // Aumentar a pontuação do jogador
+            points++;
+        }
+    }
 
     void Update()
     {
@@ -21,10 +35,10 @@ public class PlayerColeta : MonoBehaviour
     public static void MissingObject()
     {
         missingObjects++;
-        if (missingObjects >= GameOver.objetosPerdidosParaGameOver)
+        if (missingObjects >= GameControllerJC.objetosPerdidosParaGameOver)
         {
             // Game over!
-            GameOver.gameOver();
+            GameControllerJC.GameOver();
         }
     }
 }
