--- projetoIntegrador/Assets/Jogos/JogoDaColeta/Scripts/ScoreCount.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7ce1b06a6ebb3b0488bb0a6702326c87
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
