+++ projetoIntegrador/UserSettings/EditorUserSettings.asset
@@ -24,10 +24,10 @@ EditorUserSettings:
       value: 5400575753500b0855590d2045745d4413164d737e2b71662b2b1f65e7e23739
       flags: 0
     RecentlyUsedSceneGuid-6:
-      value: 515250075c0c595e5f5a5e71122159444e4e4a2f7a7d7f602f284d66b4b76661
+      value: 56520705070c585a09580e2448715e4442164f282e7c72607c2f4563b3b4606c
       flags: 0
     RecentlyUsedSceneGuid-7:
-      value: 56520705070c585a09580e2448715e4442164f282e7c72607c2f4563b3b4606c
+      value: 515250075c0c595e5f5a5e71122159444e4e4a2f7a7d7f602f284d66b4b76661
       flags: 0
     RecentlyUsedSceneGuid-8:
       value: 53075105040c0d0859085e7b457a0b441716482f2e7c20317d7b4b6abab4323d
